08-adk-HRfeedback/requirements.txt
gradio
python-dotenv
google-cloud-aiplatform
